refactor(my-diary): move functions to the actions directory
affects: @practice/my-diary, @site/app

ISSUES CLOSED: #68
